refactor: rename checkUserIdandRoomId helper
